feat:订单管理的实现

+++ src/views/product/list/AddProduct.vue
@@ -9,7 +9,7 @@
     ></el-col>
     <el-col :span="20"
       ><div class="wrapper">
-        <div class="mytitle">商品添加</div>
+        <div class="mytitle">商品{{ title }}</div>
         <el-form
           :model="ruleForm"
           :rules="rules"
@@ -33,7 +33,11 @@
             <el-input v-model="ruleForm.sellPoint" size="small"></el-input>
           </el-form-item>
           <el-form-item label="上传图片" prop="image">
-            <UpLoading @sendUrl="sendUrl" ref="uploadImg"></UpLoading>
+            <UpLoading
+              @sendUrl="sendUrl"
+              ref="uploadImg"
+              :fileList="fileList"
+            ></UpLoading>
           </el-form-item>
           <el-form-item label="商品描述" prop="descs">
             <WangEditor
@@ -63,11 +67,11 @@
             ></el-switch>
           </el-form-item>
           <el-form-item>
-            <el-button type="primary" @click="submitForm('ruleForm')"
+            <el-button v-if="title!='详情'" type="primary" @click="submitForm('ruleForm')"
               >保存</el-button
             >
-            <el-button @click="resetForm('ruleForm')">重置</el-button>
-            <el-button>取消</el-button>
+            <el-button v-if="title!='详情'" @click="resetForm('ruleForm')">重置</el-button>
+            <el-button @click="cancelPage">取消</el-button>
           </el-form-item>
         </el-form>
       </div></el-col
@@ -76,6 +80,7 @@
 </template>
 
 <script>
+import { mapState } from "vuex";
 import TreeProduct from "./TreeProduct.vue";
 import UpLoading from "./UpLoading.vue";
 import WangEditor from "./WangEditor.vue";
@@ -87,6 +92,7 @@ export default {
   },
   data() {
     return {
+      fileList: [],
       ruleForm: {
         //表单数据
         id: "",
@@ -110,7 +116,14 @@ export default {
       },
     };
   },
+  computed: {
+    ...mapState("product", ["rowData", "title"]),
+  },
   methods: {
+    /* 取消 */
+    cancelPage(){
+      this.$router.push('/product/list')
+    },
     /* 获取tree点击的数据 */
     sendTreeData(val) {
       console.log("tree", val);
@@ -142,21 +155,49 @@ export default {
             num,
             descs,
           } = this.ruleForm;
-          this.insertTbItem({
-            title,
-            image: JSON.stringify(image),
-            sellPoint,
-            price,
-            cid,
-            category,
-            num,
-            descs,
-          });
+          if (this.title === "添加") {
+            this.insertTbItem({
+              title,
+              image: JSON.stringify(image),
+              sellPoint,
+              price,
+              cid,
+              category,
+              num,
+              descs,
+            });
+          } else {
+            console.log("修改商品---", this.ruleForm, "-------------", image);
+            this.updateTbItem({
+              id,
+              title,
+              image: JSON.stringify(image),
+              sellPoint,
+              price,
+              cid,
+              category,
+              num,
+              descs,
+            });
+          }
         } else {
           console.log("error submit!!");
           return false;
         }
       });
+    },
+    /* 修改商品 */
+    async updateTbItem(params) {
+      let res = await this.$api.updateTbItem(params);
+      console.log("修改", res.data);
+      if (res.data.status === 200) {
+        this.$message({
+          message: "恭喜你，修改商品成功",
+          type: "success",
+        });
+        //跳转路由
+        this.$router.push("/product/list");
+      }
     },
     /* 添加商品 */
     async insertTbItem(params) {
@@ -173,13 +214,55 @@ export default {
     },
     /* 重置 */
     resetForm(formName) {
-      this.$refs[formName].resetFields();
-      /* 清除图片 */
-      this.$refs.uploadImg.clear();
-      /* 富文本清除 */
-      this.$refs.wangEditor.editor.clear()
-      // 第二种方法  this.$refs.wangEditor.html = "";
+      if (this.title === "添加") {
+        this.$refs[formName].resetFields();
+        /* 清除图片 */
+        this.$refs.uploadImg.clear();
+        /* 富文本清除 */
+        this.$refs.wangEditor.editor.clear();
+        // 第二种方法  this.$refs.wangEditor.html = "";
+      } else {
+        console.log("编辑界面---this.ruleForm---", this.ruleForm);
+        console.log("编辑界面---this.rowData---", this.rowData);
+        let rowData = JSON.parse(JSON.stringify(this.rowData));
+        this.ruleForm = rowData;
+        this.ruleForm.isShow = true;
+        let imgs = rowData.image; //字符串类型---需要转数组格式
+        let arr = JSON.parse(imgs); //转数组
+        console.log("---arr---", arr);
+        this.ruleForm.image = arr;
+        // console.log('arr---', arr);
+        this.fileList = [];
+        arr.forEach((ele) => {
+          this.fileList.push({ name: "", url: ele }); //fileList 回显图片--传递给UploadImg组件
+        });
+        this.$nextTick(() => {
+          this.$refs.wangEditor.html = rowData.descs;
+        });
+      }
     },
+  },
+  mounted() {
+    // 进入页面--渲染编辑的数据
+    if (this.title === "编辑" ||this.title === "详情" ) {
+      // 深拷贝
+      this.ruleForm = JSON.parse(JSON.stringify(this.rowData)); //image是字符串
+      this.ruleForm.isShow = true;
+      let imgs = this.ruleForm.image; //字符串类型
+      // 转成数组
+      let arr = JSON.parse(imgs);
+      this.ruleForm.image = arr; //让ruleForm对象里面image是数组
+      this.fileList = [];
+      arr.forEach((ele) => {
+        this.fileList.push({ name: "", url: ele });
+      });
+      console.log("fsdfes", this.fileList);
+      // 等待dom加载完毕执行 延迟执行
+      this.$nextTick(() => {
+        this.$refs.wangEditor.html = this.rowData.descs;
+      });
+      // console.log('yyyy',this.$refs.wangEditor);
+    }
   },
 };
 </script>
